worked on the request changes
